backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
Django==5.1.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
importlib_resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.1
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
